moodping-test1-shinbeom 복사본/moodping/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pymysql
python-multipart
jinja2
python-dotenv
pydantic
openai
